+++ src/Backends/Vulkan/src/frame_buffer.cpp
@@ -147,7 +147,7 @@ VulkanFrameBuffer::~VulkanFrameBuffer() noexcept = default;
 VkImageView VulkanFrameBuffer::imageView(UInt32 imageIndex) const
 {
     if (imageIndex >= m_impl->m_images.size()) [[unlikely]]
-        throw ArgumentOutOfRangeException("imageIndex", std::make_pair(0uz, m_impl->m_images.size()), imageIndex, "The frame buffer does not contain an image at index {0}.", imageIndex);
+        throw ArgumentOutOfRangeException("imageIndex", std::make_pair(0uz, m_impl->m_images.size()), static_cast<size_t>(imageIndex), "The frame buffer does not contain an image at index {0}.", imageIndex);
 
     return m_impl->m_renderTargetHandles.at(m_impl->m_images[imageIndex].get());
 }
@@ -188,7 +188,7 @@ size_t VulkanFrameBuffer::getHeight() const noexcept
 void VulkanFrameBuffer::mapRenderTarget(const RenderTarget& renderTarget, UInt32 index)
 {
     if (index >= m_impl->m_images.size()) [[unlikely]]
-        throw ArgumentOutOfRangeException("index", std::make_pair(0uz, m_impl->m_images.size()), index, "The frame buffer does not contain an image at index {0}.", index);
+        throw ArgumentOutOfRangeException("index", std::make_pair(0uz, m_impl->m_images.size()), static_cast<size_t>(index), "The frame buffer does not contain an image at index {0}.", index);
 
     if (m_impl->m_images[index]->format() != renderTarget.format()) [[unlikely]]
         LITEFX_WARNING(VULKAN_LOG, "The render target format {0} does not match the image format {1} for image {2}.", renderTarget.format(), m_impl->m_images[index]->format(), index);
@@ -219,7 +219,7 @@ Enumerable<const IVulkanImage*> VulkanFrameBuffer::images() const noexcept
 const IVulkanImage& VulkanFrameBuffer::image(UInt32 index) const
 {
     if (index >= m_impl->m_images.size())
-        throw ArgumentOutOfRangeException("index", std::make_pair(0uz, m_impl->m_images.size()), index, "The frame buffer does not contain an image at index {0}.", index);
+        throw ArgumentOutOfRangeException("index", std::make_pair(0uz, m_impl->m_images.size()), static_cast<size_t>(index), "The frame buffer does not contain an image at index {0}.", index);
 
     return *m_impl->m_images[index];
 }
